--- src/cart/cart.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CartController } from './cart.controller';
-
-describe('CartController', () => {
-  let controller: CartController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [CartController],
-    }).compile();
-
-    controller = module.get<CartController>(CartController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
